Project restructur.
Generic path to containers add.
More general structure.
